+++ Samples/SampleChatAppWithShare/SampleChatAppWithShare/SampleChatAppWithShare.exe.manifest
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<assembly manifestVersion="1.0" xmlns="urn:schemas-microsoft-com:asm.v1">
+  <assemblyIdentity version="1.0.0.0" name="SampleChatAppWithShare.app"/>
+  <msix xmlns="urn:schemas-microsoft-com:msix.v1"
+          publisher="CN=ravishroshan"
+          packageName="Weixin"
+          applicationId="Weixin"
+        />
+</assembly>
